[ENGOPS-3025] update parent pom to 7.1-SNAPSHOT & compile with jdk 1.8

+++ api/core/src/test/java/org/pentaho/profiling/api/metrics/field/DataSourceFieldValueTest.java
@@ -62,7 +62,7 @@ public class DataSourceFieldValueTest extends BeanTester {
     int metadataVal = 50;
     val.setFieldMetatdata( metadataKey, metadataVal );
     assertNotNull( val.getFieldMetadata( metadataKey ) );
-    assertEquals( metadataVal, val.getFieldMetadata( metadataKey ) );
+    assertEquals( metadataVal, (int) val.getFieldMetadata( metadataKey ) );
     val.clearFieldMetadata();
     assertNull( val.getFieldMetadata( metadataKey ) );
   }
